Small changes related to Typo & shifting logic of parking fee

+++ src/com/lld/examples/strategies/FeeStrategy.java
@@ -1,5 +1,5 @@
 package com.lld.examples.strategies;
 
 public interface FeeStrategy {
-	public int calculateParkingFee(int hours);
+	public int calculateParkingFee(long hours);
 }
